src/requirements.txt
absl-py==0.1.10
astor==0.6.2
backports-abc==0.5
backports.functools-lru-cache==1.4
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0rc1
beautifulsoup4==4.6.0
bleach==1.5.0
bs4==0.0.1
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
cycler==0.10.0
Cython==0.28.1
decorator==4.2.1
enum34==1.1.6
Flask==0.12.2
funcsigs==1.0.2
futures==3.1.1
gast==0.2.0
glob2==0.6
grpc==0.3.post19
grpc-wrapper==0.4
grpcio==1.10.0
grpcio-tools==1.7.0
h5py==2.7.1
html5lib==0.9999999
idna==2.6
imageio==2.3.0
ipdb==0.11
ipykernel==4.8.2
ipython==5.5.0
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.10
jupyter-client==5.2.3
jupyter-core==4.4.0
Keras==2.1.5
lxml==4.1.1
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==2.1.1
mock==2.0.0
moviepy==0.2.3.4
msgpack-python==0.5.6
munch==2.3.1
natsort==5.1.1
nltk==3.2.5
numpy==1.13.3
opencv-python==3.4.0.12
pandas==0.22.0
pathlib2==2.3.0
pbr==3.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
protobuf==3.5.0.post1
ptyprocess==0.5.2
pygame==1.9.3
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
pyzmq==17.1.0
requests==2.18.4
scandir==1.7
scikit-learn==0.19.2
scipy==1.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sk-video==1.1.10
stn==1.0.1
subprocess32==3.2.7
telegram==0.0.1
tensorboard==1.6.0
tensorflow-gpu==1.4.0rc1
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
tornado==5.1
tqdm==4.23.3
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
